+++ tests/repr_test.py
@@ -14,7 +14,7 @@
 
 from __future__ import annotations
 from grscheller.fp.woException import MB, XOR
-from grscheller.datastructures.split_ends import SplitEnd
+from grscheller.datastructures.stacks import SplitEnd
 from grscheller.datastructures.queues import DoubleQueue
 from grscheller.datastructures.queues import FIFOQueue
 from grscheller.datastructures.queues import LIFOQueue