+++ CHANGELOG.md
@@ -20,6 +20,10 @@ PyPI grscheller.datastructures project.
 
 ## Releases and other important milestones
 
+### Version 0.21.1.0 - mostly docstring updates 2024-09-17
+
+* renamed module split\_ends -> stacks
+
 ### Version 0.21.0 - PyPI release date 2024-08-20
 
 * got back to a state maintainer is happy with